bump offset to uint64

+++ packfile/packfile_test.go
@@ -40,7 +40,7 @@ func TestPackFile(t *testing.T) {
 	if p.Footer.Count != 2 {
 		t.Fatalf("Expected Footer.Count to be 2 but got %d", p.Footer.Count)
 	}
-	if p.Footer.IndexOffset != uint32(len(p.Blobs)) {
+	if p.Footer.IndexOffset != uint64(len(p.Blobs)) {
 		t.Fatalf("Expected Footer.Length to be %d but got %d", len(p.Blobs), p.Footer.IndexOffset)
 	}
 }
@@ -155,7 +155,7 @@ func TestPackFileSerializeFooter(t *testing.T) {
 
 	require.Equal(t, p2.Count, uint32(2), "Expected 2 blobs but got %d", uint32(p2.Count))
 
-	require.Equal(t, p2.IndexOffset, uint32(len(chunk1)+len(chunk2)), "Expected IndexOffset to be %d but got %d", len(chunk1)+len(chunk2), p2.IndexOffset)
+	require.Equal(t, p2.IndexOffset, uint64(len(chunk1)+len(chunk2)), "Expected IndexOffset to be %d but got %d", len(chunk1)+len(chunk2), p2.IndexOffset)
 }
 
 func TestPackFileSerializeData(t *testing.T) {
@@ -183,5 +183,5 @@ func TestPackFileSerializeData(t *testing.T) {
 func TestDefaultConfiguration(t *testing.T) {
 	c := DefaultConfiguration()
 
-	require.Equal(t, c.MaxSize, uint32(20971520))
+	require.Equal(t, c.MaxSize, uint64(20971520))
 }
